제휴 상세페이지 작성 중

+++ sas-back/src/main/java/kr/co/sas/store/model/dto/StoreDTO.java
@@ -33,8 +33,8 @@ public class StoreDTO {
 	private int deposit;
 	private int registType;
 	private String storeIntroduce;
-	private int mapX;
-	private int mapY;
+	private double mapX;
+	private double mapY;
 	private int type;
 	private int servicePrice;
 	private String storeRequestDate;
